Projet_Rakuten_docker/Conteneur_BDD_utilisateur/requirements.txt
fastapi
uvicorn
passlib
python-multipart
PyJWT
pytest
httpx
